Use git client 2.4.4 - better protect temporary files
Temporary files used by the command line git implementation may contain
sensitive information (like user names, passwords, private keys,
or passphrases).  Earlier releases of the git client used the default
permissions and the default system temporary directory to store those
files while they were used.

This release of the git client prefers to create those temporary files
in a temporary workspace associated with the job, if a workspace
is available.  If a workspace is not available, it uses the system
temporary directory.  In either case, it also protects the files so that
they are only readable by the owner, and if they are executable scripts,
they are only executable by the owner.