Revert "Add vcvarsall plugin"
Does not seem to detect Visual Studio 2017 Community Edition on the three
machines I tested.  I enabled the checkbox, and pressed the Save button
for that agent and the web page spun while waiting for the operation
to complete.  If I open that same agent web page from another browser,
the results seem to not have been saved yet.

This reverts commit 1acd1994e7b9c0914618412ed779a3227ede6f2e.