Add [JENKINS-29603] verification jobs
The job with a single branch shows the bug.

The job with multiple branches does not show the bug.

There were several other interesting results that need more investigation
in the comparison between the jenkins-bugs local branch and the
jenkins-bugs no-local-branch.

Need a jenkins-bugs from a GitHub branch source as well.