Add JENKINS-37263 separate check job
Used the job to check a customer question on polling.  It appears that
polling (at least with this job definition) does not detect changes
when I use some of the alternate forms of branch name as listed in the
online help.

This job is distinct because it honors the refspec on clone, and it uses
the features/* syntax in the refspec.  I was expecting that would match
all branches under features when polling, but it did not when I used
the branch name as:

    refs/heads/features/*
    remotes/origin/features/*
    refs/remotes/origin/features/*

It worked as expected when I used:

    */features/*
    origin/features/*