Use correct path to maven on master
Relies on tool installer location not changing, but that seems no
worse than the ugly requirement that started all this.  A maven tool
needs to be available on each agent based on what I understand about
buildPlugin.  The buildPlugin shared library does not seem to tie
itself to any label which would require maven on the agent.