Disallow jdk10 test on master agent
Master agents running Alpine Linux can't use the same JDK bundle as
other masters.  Rather than fight with that problem at the moment,
exclude jdk10 tests from the master agent.

On the positive side, this motivates a faster transition to use OpenJDK
as provided with the operating system distribution rather than the Oracle
JDK builds.

Something similar will be needed in other areas as more Alpine Linux
surprises are discovered.

The 'cmake' tool in this repo is using a different technique.  It adds a
'cmake-locally-installed' tool which will use whatever cmake version is
installed on the local operating system.  That moves the responsibility
for repeatable configuration from the job definition to the agent
definition but has the benefit that the tools are then installed and
maintained by agent definition commands (operating system package
management) rather than job definitions.