Retain many more build chooser builds for BuildData fix
Build chooser needs enough history to record each branch that is being
built.  Since BuildData no longer retains the entire history of every
preceding build in each job, we need to keep a longer history.

In this specific case, there needs to be enough history to record the
build status of all the branches selected by the BuildChooser.