Set JENKINS-33827 daysToKeep to -1
Was previously using 0 as the value.  The user interface indicates that
0 is not a positive integer. The history seems to retain only the most
recent build when the daysToKeep is set to 0.

That change was sufficient to resolve several of the build chooser
infinite build problems with the git plugin change that removes the
BuildData bloat.  However, it does not seem to have removed all the
cases of build chooser infinite builds.

Also increases the builds to keep to 193 in case the number of matching
branches continues to increase.