Use analysis core plugin 1.96 (special case)
The findbugs plugin is still referenced by the stable-3.9 branch of
the git plugin.  Since the git plugin stable-3.9 branch is expected to
continue for a long time, this plugin is needed so that the findbugs
step can be used in that plugin branch.

The stable-3.10 and master branches of the git plugin have switched to
spotbugs and do not need the findbugs plugin.  This plugin can be
removed when the stable-3.9 branch has been idle long enough to not be
relevant, or when the findbugs call in the stable-3.9 Jenkinsfile has
been removed.