Don't use CentOS 6 - fails with 'did you run git update-server-info on the server'
The failure message is specific to command line git 1.7.1.  Since that
version of command line git is not officially supported by the git plugin,
this seems like a safe case where the test job can be reasonably defined
to avoid the known failure with an unsupported command line git version.