Add external monitor job plugin 1.7
Unclear why this plugin is now being installed when it was not being
installed previously. I don't need the plugin and did not request the
plugin, yet some update to another plugin or to Jenkins core has now
caused it to be installed.

Rather than have the update manager regularly show there is a plugin
to be updated, update it and don't worry about it not being needed in
this environment.