Use windows-slaves 1.4
Unclear to me why this plugin is now being installed.  I don't need
it.  It does not work in any case that is useful to me. I'm not aware
of anyone that has it working in a useful case.

Some other plugin update seems to have included this plugin as a
dependency, even though I'm confident that it is an optional
dependency.