Check JENKINS-49757 with a refspec of non-default values
The default refspec maps refs/heads/* to refs/remotes/origin/* .
It does not map the GitHub special refs/pull/* to anything.

This sample job maps refs/pull/553 to refs/remotes/origin/pull/553
by altering the refspec.  In the case where the second fetch is used,
that altered refspec will be used.  If the seceond fetch is eliminated,
the mapping from refs/pull/553 to refs/remotes/origin/pull/553 will
not happen.

The job uses the xshell plugin to run `git branch -a`.

The job uses the text-finder plugin to search the console log for
the expected output from the `git branch -a` command.