[JENKINS-62767] JGit should only be included in git-client plugin
Including jgit in any other plugin risks that the JGit version provided by
the other plugin is the incorrect version.  JGit is provided intentionally
as part of the git-client plugin API.  When other plugins include JGit,
they risk violating the contract that git client plugin has with the
rest of the Jenkins plugins.