Windows git tools match specific linux git versions
Agents are labeled with the two element git version that describes the
major and minor release of git that is installed.  For example, Ubuntu
18 is labeled with 'git-2.17'.

A git tool is defined for 'git-2.17.1' which has two installers, one
for agents with the label 'git-2.17' and the other for agents with the
label 'windows'.  Agents with the label 'git-2.17' are assumed to
already have command line git 2.17 installed at /usr/bin/git.  Agents
with the label 'windows' will download the tool from
https://home.markwaite.net/~mwaite/git/

The technique will also allow me to assign a specific git patch
version for other operating systems like FreeBSD and OpenBSD, where
git is installed at /usr/local/bin/git.

Will need to remember to only label agents with the major and minor
versions of git so that the patch version is always a reference to a
tool and not to a label.