Check JENKINS-63785 in various ways
Second fetch is being skipped even if it contains a parameter that
must be expanded.  I would assume that since the plugin does not
attempt to determine the value of the expanded string, it cannot
reliably decide to ignore the second fetch when the refspec is not
honored on the first fetch and the refspec contains a variable that
needs to be expanded.