Use git client plugin 3.5.2-SNAPSHOT (private-e7f34e26-mwaite)
Requires POST for the doCheckHome for tidiness and to avoid a scanner
warning.  The check is entirely on the local file system, so this is
not a security risk, but rather an improvement to acknowledge that
there is no konwn use of this method outside POST from a Jenkins
configuration page.