Allow any local agent to ruyn JENKINS-26582-launcher
The job needs curl access to the Jenkins URL but does not require
anything more than that.  No need to require it run on the Jenkins
controller. It is enough to have it run on an agent that has HTTP access
to the controller.

Also reduced the footprint of the cloned repository by defining a refspec
that only retrieves one branch.