Remove controller executor disable from jobs
The controller needs to execute certain special case jobs (like these).
However, we don't want execution to be allowed on the controller in
general.  There is a periodic job defined that enables executors on
the controller, sleeps for a brief time, then disables executors on
the controller.  If there are jobs queued to run on the controller,
they will begin execution soon after the executors are available.

Now covered in the job that periodically enables rather than having
multiple locations that disable the executors on the controller after
they have been enabled.

The enable job now delays for a fixed period and then disables the
executors on the controller.