Remove automatic labels from built-in agent
Reduce risk of platform targeted jobs running on built-in agent.  Jobs
that were designated for `Debian` were being run on the built-in agent
because the agent is configured to accept jobs that are labeled for it.
Remove the simplified operating system specific label so that the
built-in node is not labeled with more generic operating system labels.

It is not a generic agent, so it should not have general purpose labels
applied.