Use git plugin 5.0.0-rc5167.b_a_089575e15d
Includes the submodule timeout default value fix, scmGit as the
preferred symbol, lfs as the symbol to enable LFS, and gitSCM as an
optional symbol.

Still do not have a solution to the configuration as code warning that
gitSCM is deprecated when I can't find the changes in the configuration
file that will prevent the message.