Switch to accept first connection strategy
The ssh host key is reported as changed even though I can't determine
the source of the incorrect host key.  I had updated it and still it
reported that the key changed.  I adjusted the list of known hosts and
still it reported that the only line in the known hosts file was showing
that the host key had changed.

This won't be smooth for static agents, since they remember the previous
host key in their known hosts file.