Fix 'vm' implied label - all are VM's except alpine
In my configuration, all the agents that run in a container use the
Alpine operating system.  That is not a general purpose restriction,
but happens to work in my configuration at the current time.

If more operating system types are added as container agents, then
this definition of 'vm' will need to be handled in a different way.
For example, if container agents and virtual machine agents can both run
Alpine Linux, then some other way is needed to identify a container agent.